refactor Athlete and Performance Uploads into separate components
